# Fix: make partition work for forward ranges