The current `partition<>` implementation, the one or two input collections passed are assumed internally to be random access containers (to be precise, it is required that the collection's `operator[]` is defined).
In fact, this is requirement is a bit too strict. `partition<>` can easily work even for forward ranges, and this is the focus of this PR.

In this PR, `partition<>` has been modified so as to work for forward ranges. This eliminates among other things the need for the range to define `operator[]`. The idea behind the re-implementation is as follows: instead of storing and keeping track of indices of collection items (these are the indices that are used as arguments in `operator[]` calls), iterators to collection items are stored and manipulated. The `index_vector_type` is consistently replaced by a vector of iterators (possibly of two different types if the two input collections are different). Accessing the collection items in this context simply amounts to de-referencing the iterators stored in the iterator vectors. The only operators used with respect to the iterators of the input collections are `operator++` and `operator*`.

On the practical side: this allows to use partition with ranges which are not random access. Two such cases are ranges coming from BG's `point_iterator<>` and `segment_iterator<>` which are bidirectional iterators.

All algorithm unit tests pass with the re-designed `partition<>`.