@barendgehrels 2015-03-06
I wonder why you call this "Fix". It is obviously not a fix, but an improvement, making it more generi