@barendgehrels 2015-03-06
I also wonder why you call this "reimplement" and "redesign". Anyway, does not matter, just wondering - it is just changing a part of the implementation