@mkaravel 2015-03-06
* About "Fix": partly lack of creativity in the naming, and partly that it "fixes" the "shortcoming" of working for random access ranges only.
* About 1.58: no problem: my usage case is difference and intersection of pointlike/linear geometries, so it is not really needed for 1.58.
* About "reimplement" and "redesign": partly lack of creativity and imagination for naming what I did, and partly the fact that I use iterators instead of indices (which in my mind is a design choice). Honestly, I never intended to mean that this is my/new design or my/new implementation; it seemed more natural to call what I did a "redesign" or a "reimplementation" rather than a "change". But this is just my own opinion, and I do not have strong feelings about it. What is important for me here is the added functionality.