@barendgehrels 2015-03-06
OK, thanks for the explanations. Good to hear you want to use it for point like differences.