@mkaravel 2015-03-06
It is for pointlike-linear differences. Pointlike-pointlike are already implemented in a different way (sorting and binary search; I did benchmarks then, and it was faster that way compared to using partition).