@barendgehrels 2015-03-06
Sure, pointlike and linear. We still have to make partition so fast that it can compare to std::sort ....