@mkaravel 2015-03-09
## **_Performance comparison_**

### **Setup**
I tried the following scenario: I considered random points in a `[0,1] x [0,1]` square, seen as multipoints. For these multipoints I called partition. The visitor passed is not doing something very meaningful, but still I wanted it to do something to avoid any possible optimizations by the compiler. What it does is that for the points in each bottom-most level box, it computes the minimum distance among them. I considered running the old and new implementation of partition either for a single multipoint (one-collection version of partition) or for two multipoints (two-collections version of partition).

See the following files (the partition implementations were given different names and put in different namespace so that I can use both of them in the same program):
* Comparison/timing program: https://github.com/mkaravel/geometry/blob/performance-partition/performance/algorithms/partition/multipoint_multipoint.cpp
* Old partition implementation: https://github.com/mkaravel/geometry/blob/performance-partition/performance/algorithms/partition/partition_old.hpp
* New partition implementation: https://github.com/mkaravel/geometry/blob/performance-partition/performance/algorithms/partition/partition_new.hpp

### **Compilation**
The comparison program was compiled with `clang++` version 3.4 with the flags `-O2` and `-DNDEBUG`.

### **Timings**
The following two tables show the elapsed times for the old and new partition implementation with respect to the size of the multipoint(s) passed.
Times are in seconds. `Boost.Timer` has been used for performing the timings. Smallest time is shown in bold-face.

| *Multipoint size* | *Old* | *New* |
|:-------------:|--------:|--------:|
| 10^3 | 0.000618 | **0.000585** |
| 10^4 | 0.002773 | **0.002450** |
| 10^5 | **0.035126** | 0.036622 |
| 10^6 | 0.431059 | **0.394527** |
| 10^7 | 5.861435 | **5.680503** |


| *Multipoint  # 1 size* | *Multipoint  # 2 size* | *Old* | *New* |
|:-------------:|:-------------:| --------:|--------:|
| 10^3 | 10^3 | 0.001106 | **0.000997** |
| 10^4 | 10^4 | 0.005468 | **0.005185** |
| 10^5 | 10^5 | 0.076411 | **0.065967** |
| 10^6 | 10^6 | 0.849864 | **0.813731** |
| 10^7 | 10^7 | 12.670377 | **12.326983** |