@mkaravel 2015-03-09
My understanding is that there is no show-stopper for this PR.

Since we have decided to leave it for boost 1.59, please just comment that the PR is indeed okay, and we can merge once our develop branch is open for 1.59