@barendgehrels 2015-03-09
The timings look convincing, thanks for this. We already earlier noticed that changing from index operator to iterators is measurably faster.
So yes, we can merge it for 1.59